ext4: Re-organize files
Moves everything into its own file, and gets rid of the horrible
ext4_ll. ext4.cpp now takes care of basic implementations of the VFS
ops.
